`titleFormat` added to the `EntityMeta` GraphQL type
The fields constains a list of entity fields to be used to compose both title and subtitles for an entity; useful while listing them or an interface for a particular entity.

For now the `titleFormat` is hard-coded with information for the Example Data Domain, but later it should be evaluated by inspecting the data schema / data values and allow for user overrides.