Entity Type Introspection [WIP]
The entity types are no longer hard coded into the GQL response. Now they are derived from the domain DB.

The derivation process is not perfect, so overrides will be allowed to be made by users. Therefore a meta DB was introduced (memory for now, just like domain DB) and will persist already derived metas (not yet done).

Fetching the actual data for a particular entity type is still hard coded to the data domain (there both `users()` and `features()` coded to fetch them from the DB). This should probably be the next issue to be addressed.

As of custom, lots of `// TODO` were addded. Not of custom, the suffix `WIP` (`// TODO WIP`) was used to indicate issues that should be considered part of the current story.