Canceling `code` field deletion
In the end, the `code` field is useful to reference an entity in logs and errors. It is like a name, but shorter.

Coupling it to the table name is the wrong thing. The table may be used as the entity code, but the user may choose to override it.