Decoupling `code` from `table` and `column`
The `code` field was being used to serve as both the internal DB binding (should be hidden from user) and readable programatic reference to entity and field types.

At first, I attempted removing the `code` field altogether, but it proved useful to assemble entity JSON (field codes used as object keys) and reference entity/field types by a readable name in logs and errors (instead of a number). So the `code` field must remain.

The problem is that, as it was implemented, there would be no way to hide database details from the API interface. Decoupling `code` from `table`/`field` allows for it to be hidden.

Note that the current introspection algorithm does use the table/column nas as code, but it now can be overriden by the user.