Extracting `metaDB` and `domainDB` creation from `DataDomain`
Previously, `DataDomain` was responsible for connecting, creating and populating both databases used. This is a bit too much for it and it should receive connections to these databases and expect them to be already created.

The `domainDB` creation/population was particularly troublesome; the `DataDomain` is expected to inspect the domain schema, not create it. Note that the problem was just moved to `main.ts`, since it is now creating and populating the `domainDB` with the example data domain schema/data in case `CREATE_EXAMPLE_DATA_DOMAIN` is `TRUE`.

Regarding `metaDB`, it is ok for `DataDomain` to populate it with the inferred metadata. `metaDB`'s data belongs and is controlled by `DataDomain`.

Also:
- Some TODOs have been moved to Trello.